requirements.txt
kokoro-onnx
gradio 
wget
soundfile
ollama
requests
whisper-openai
pyaudio